Cleaned up database diff test
