+++ lib/databaseDescribe.js
