+++ lib/databaseDescribe.sql
