+++ lib/databaseDiff.js
@@ -9,9 +9,8 @@ const colors = require('colors');
 const _ = require('lodash');
 const jsdiff = require('diff');
 
-const sqldb = require('../lib/sqldb');
 const sqlLoader = require('../lib/sql-loader');
-const describe = require('../tests/databaseDescribe');
+const describe = require('./databaseDescribe');
 
 module.exports = {};
